fix(docker): Don't load native on alpine image, fix #6510
