+++ packages/cubejs-backend-shared/src/index.ts
@@ -18,4 +18,5 @@ export * from './cli';
 export * from './proxy';
 export * from './time';
 export * from './process';
+export * from './platform';
 export * from './FileRepository';
