+++ rust/cubestore/js-wrapper/src/utils.ts
@@ -1,54 +1,5 @@
-import { spawnSync } from 'child_process';
 import process from 'process';
-import { displayCLIWarning, internalExceptions } from '@cubejs-backend/shared';
-
-export function detectLibc() {
-  if (process.platform !== 'linux') {
-    throw new Error('Unable to detect libc on not linux os');
-  }
-
-  try {
-    const { status } = spawnSync('getconf', ['GNU_LIBC_VERSION'], {
-      encoding: 'utf8',
-      // Using pipe to protect unexpect STDERR output
-      stdio: 'pipe'
-    });
-    if (status === 0) {
-      return 'gnu';
-    }
-  } catch (e: any) {
-    internalExceptions(e);
-  }
-
-  {
-    const { status, stdout, stderr } = spawnSync('ldd', ['--version'], {
-      encoding: 'utf8',
-      // Using pipe to protect unexpect STDERR output
-      stdio: 'pipe',
-    });
-    if (status === 0) {
-      if (stdout.includes('musl')) {
-        return 'musl';
-      }
-
-      if (stdout.includes('gnu')) {
-        return 'gnu';
-      }
-    } else {
-      if (stderr.includes('musl')) {
-        return 'musl';
-      }
-
-      if (stderr.includes('gnu')) {
-        return 'gnu';
-      }
-    }
-  }
-
-  displayCLIWarning('Unable to detect what host library is used as libc, continue with gnu');
-
-  return 'gnu';
-}
+import { displayCLIWarning, internalExceptions, detectLibc } from '@cubejs-backend/shared';
 
 export function getTarget(): string {
   if (process.arch === 'x64') {
